refactor: split technology_panel slot/chip widgets into sibling file (Refs #3510)
Extract ResearchedTechChip, TechSectionHeading, ResearchSlotCard,
LockedResearchSlotCard and their private slot-card chrome into
technology_panel_widgets.dart, re-exported from technology_panel.dart so
existing imports keep resolving. Brings technology_panel.dart back under
the 700-line repo.game_widgets_file_size cap that was failing CI.
